requirements.txt
openai
dotenv
diskcache
sqlglot
pandas
PyMySQL
ollama
pyyaml
